requirements.txt
fastapi
uvicorn
supabase
pydantic
pillow
numpy